requirements.txt
# Core dependencies
PyMuPDF==1.23.8
pytesseract==0.3.10
Pillow==10.1.0
opencv-python==4.8.1.78
numpy==1.24.3

# ML/NLP
sentence-transformers==2.2.2
transformers==4.35.2
torch==2.1.0
spacy==3.7.2
scikit-learn==1.3.2

# Search
faiss-cpu==1.7.4
rank-bm25==0.2.2

# Graph/Network
networkx==3.2.1

# API/Server
fastapi==0.104.1
uvicorn==0.24.0
websockets==12.0
pydantic==2.5.0
aiofiles==23.2.1

# Storage/Cache
redis==5.0.1
aioredis==2.0.1
pyyaml==6.0.1

# Monitoring
prometheus-client==0.19.0

# Utils
tqdm==4.66.1
pandas==2.1.3
python-multipart==0.0.6

# Optional but recommended OCR engines
easyocr==1.7.0